Separated order and payment models
